Add files via upload
This notebook focuses on demonstrating fundamental string operations in Python. This educational notebook serves as an introductory tutorial to handling and manipulating strings, which is a vital skill in many areas of programming, particularly in data cleaning and preparation phases of data analysis.
The notebook presents a series of string operations step-by-step:
Repetition: It demonstrates how to repeat a string multiple times using the multiplication operator (*), which is useful for generating repeated data or patterns.
Indexing: It shows how to access characters in a string by their index using square brackets ([]), which is fundamental for slicing and dicing data.
Length: It uses the len() function to calculate the number of characters in a string, an essential function for assessing the size of textual data.
Find: It explains how to locate a substring within another string using the find() method, which returns the position of the substring or -1 if the substring is not found. This operation is crucial for text parsing and data validation tasks.