Add files via upload
This notebook focuses on the preprocessing and analysis of textual data using Python. This notebook demonstrates essential natural language processing (NLP) techniques using popular libraries such as pandas for data handling and NLTK for text processing.

The notebook begins by loading a dataset named "emotion_sentimen_dataset.csv" into a pandas DataFrame. This dataset presumably contains textual data related to emotions or sentiments, which is ideal for NLP tasks. The initial steps include reading the data from a CSV file and displaying the first few entries to provide insight into the dataset's structure and content.

Following data loading, the notebook emphasizes the importance of preprocessing in text analysis. It starts by downloading and using a set of English stopwords from the NLTK library. Stopwords are commonly used words (such as "the", "is", "in") that are typically removed from the text data because they offer no significant value in NLP tasks aimed at understanding sentiments or emotions.

The core of the notebook is the development of a text preprocessing pipeline, which includes:

Lowercasing the text to ensure that the analysis is case-insensitive.
Tokenization, which involves breaking down text into words or meaningful tokens using regular expressions.
Removing stopwords from the tokens to focus on more meaningful words.
Each step in this pipeline is implemented as a function, and these functions are sequentially applied to the text data within the DataFrame. The processed tokens are then added as a new column in the DataFrame, allowing for easy access and analysis in subsequent steps.